fix(utimens): store the epoch bits, and widen seconds to int64
G2 taught the READER that an ext4 timestamp is a signed 32-bit base
plus two epoch bits in the matching `*_extra` field. The WRITER was
left behind: `apply_utimens` took a `u32`, wrote it into the base and
left the epoch bits zero.

So the two halves disagreed. Setting an mtime of 2046 stored a base
that is negative as an i32, and the new signed decoder read it back
as a date in 1909 -- 136 years early, from a call that returned 0.

The setter now takes `i64` seconds and encodes them the way the
kernel does. Three things fall out of the widening:

- Pre-1970 times are expressible at all. A `u32` could not name one.
- `u32::MAX` is no longer a spare bit pattern -- it is an ordinary
  date in 2106 -- so the "leave this one alone" sentinel moves to
  `FS_EXT4_TIME_OMIT` (`INT64_MIN`).
- A time outside what the format can hold (roughly 1901..2446), or
  one needing epoch bits on a 128-byte inode with no `*_extra` field,
  is refused with EINVAL. The check runs before any byte is written,
  so a refused call leaves the inode exactly as it was.

`encode_extra_time` moves out of the test module and into `inode.rs`
beside its decoder. It existed only as a test-local twin, so the
round-trip tests were checking two copies of the encoder against each
other while the shipped code had none.

The header spells the sentinel `static const int64_t`, not `#define`:
Swift's C importer takes simple literal macros only, and silently
drops `#define FS_EXT4_TIME_OMIT INT64_MIN`. Verified by typecheck.

Also derives `the_other_ro_compat_bits_still_mount` from
SUPPORTED_RO_COMPAT rather than hand-listing the bits. The list had
omitted GDT_CSUM since the first release, so a regression rejecting a
supported feature would have passed. Mutation-checked: rejecting
GDT_CSUM now fails the test, and did not before.

Tests: +3 (post-2038 round trip, pre-1970 round trip, out-of-range
refusal). 111 suites green, clippy clean.

+++ src/capi.rs
@@ -2175,19 +2175,28 @@ pub unsafe extern "C" fn fs_ext4_set_flags(
 }
 
 /// Set the access + modification times on `path`. Each `*_sec` is the
-/// POSIX seconds-since-epoch; pass `u32::MAX` (0xFFFF_FFFF) to leave a
-/// given pair unchanged. `*_nsec` are the sub-second nanoseconds (written
-/// only when i_extra_isize covers them). Bumps i_ctime.
+/// POSIX seconds-since-epoch, signed and 64-bit to match what the format
+/// stores; pass `FS_EXT4_TIME_OMIT` (`INT64_MIN`) to leave a given pair
+/// unchanged. `*_nsec` are the sub-second nanoseconds (written only when
+/// i_extra_isize covers them). Bumps i_ctime.
+///
+/// Widened from `uint32_t` in 0.5.0. The old type could not express a
+/// pre-1970 date, and its `UINT32_MAX` sentinel collided with a real
+/// date in 2106; more to the point, a caller passing a post-2038 value
+/// got it stored without the epoch-extension bits and read it back 136
+/// years early.
 ///
 /// Returns 0 on success, -1 on failure with details in
-/// `fs_ext4_last_error`.
+/// `fs_ext4_last_error`. Fails with EINVAL for a time outside the range
+/// ext4 can store, or one needing the epoch bits on an inode too small
+/// to carry an `*_extra` field.
 #[no_mangle]
 pub unsafe extern "C" fn fs_ext4_utimens(
     fs: *mut fs_ext4_fs_t,
     path: *const c_char,
-    atime_sec: u32,
+    atime_sec: i64,
     atime_nsec: u32,
-    mtime_sec: u32,
+    mtime_sec: i64,
     mtime_nsec: u32,
 ) -> c_int {
     ffi_guard(
